split keccak circuit into mods
